enh(broker/muxer_filter): new filter mechanism in broker
* MON-12324 add monitoring stream filter

* MON-12324 add bam graphite rrd filters

* MON-12324 add unified_sql filter

* MON-12324 an error message is logged if configration filters mandatory messages

* MON-12324 use muxer_filter for user filters

* MON-12324 add test

* MON-12324 fix link issue

* MON-12324 fix test

* MON-12324 move multiplexing and new filter policy

* enh(broker/io::endpoint): it is directly constructed with stream filters

* fix(broker/muxer_filter): internal events should be allowed everytimes

* fix(broker/processing): two attributes forgotten in processing/acceptor

* fix(broker/endpoint): reference is a bad idea here

* cleanup(broker/muxer): muxer filters are simplified

* fix(broker/unified_sql): missing constexpr to compile with clang

* cleanup(broker): compile commands in broker/engine removed

* fix(broker/bam): an issue with new connector

* fix(tests/muxer_filter): an error in a comparison is fixed

* cleanup(broker/config): coding style applied

* cleanup(broker): comments modified

* fix(broker/feeder): argument passed by reference

* fix(broker/muxer): arguments passed by reference

* fix(broker/muxer): unused attributes removed

* fix(tests): bloody root_centreon

* enh(tests): new tests on filters

* fix(tests): regression on tests

* enh(tests): missing file added

* cleanup(broker): little things forgotten

* fix(tests): some python code improved

* fix(tests:): test on filters more accurate

---------


REFS: MON-12324

+++ broker/neb/src/callbacks.cc
@@ -2018,7 +2018,8 @@ int neb::callback_host_status(int callback_type, void* data) {
             static_cast<neb::pb_acknowledgement*>(it->second.get());
         if (!(!host_status->current_state  // !(OK or (normal ack and NOK))
               || (!a->obj().sticky() &&
-                  host_status->current_state != a->obj().state()))) {
+                  host_status->current_state !=
+                      static_cast<short>(a->obj().state())))) {
           a->mut_obj().set_deletion_time(time(nullptr));
           gl_publisher.write(std::move(it->second));
         }
@@ -2127,7 +2128,8 @@ int neb::callback_pb_host_status(int callback_type, void* data) noexcept {
       neb::pb_acknowledgement* a =
           static_cast<neb::pb_acknowledgement*>(it->second.get());
       if (!(!hscr.state()  // !(OK or (normal ack and NOK))
-            || (!a->obj().sticky() && hscr.state() != a->obj().state()))) {
+            || (!a->obj().sticky() &&
+                hscr.state() != static_cast<short>(a->obj().state())))) {
         a->mut_obj().set_deletion_time(time(nullptr));
         gl_publisher.write(std::move(it->second));
       }
@@ -3395,7 +3397,8 @@ int32_t neb::callback_pb_service_status(int callback_type [[maybe_unused]],
       neb::pb_acknowledgement* a =
           static_cast<neb::pb_acknowledgement*>(it->second.get());
       if (!(!sscr.state()  // !(OK or (normal ack and NOK))
-            || (!a->obj().sticky() && sscr.state() != a->obj().state()))) {
+            || (!a->obj().sticky() &&
+                static_cast<uint32_t>(sscr.state()) != a->obj().state()))) {
         a->mut_obj().set_deletion_time(time(nullptr));
         gl_publisher.write(std::move(it->second));
       }
@@ -3528,7 +3531,8 @@ int neb::callback_service_status(int callback_type, void* data) {
             static_cast<neb::pb_acknowledgement*>(it->second.get());
         if (!(!service_status->current_state  // !(OK or (normal ack and NOK))
               || (!a->obj().sticky() &&
-                  service_status->current_state != a->obj().state()))) {
+                  service_status->current_state !=
+                      static_cast<short>(a->obj().state())))) {
           a->mut_obj().set_deletion_time(time(nullptr));
           gl_publisher.write(std::move(it->second));
         }
